+++ tests/FamilyHub.IdentityServerHost.UI.UnitTests/MyAccount/WhenUsingChangeEmail.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FamilyHub.IdentityServerHost.UI.UnitTests.MyAccount;
+
+public class WhenUsingChangeEmail
+{
+}
